+++ packages/engine/Source/Scene/ITwinData.js
@@ -7,7 +7,6 @@ import Check from "../Core/Check.js";
 import KmlDataSource from "../DataSources/KmlDataSource.js";
 import GeoJsonDataSource from "../DataSources/GeoJsonDataSource.js";
 import DeveloperError from "../Core/DeveloperError.js";
-import deprecationWarning from "../Core/deprecationWarning.js";
 
 /**
  * Methods for loading iTwin platform data into CesiumJS
@@ -45,20 +44,11 @@ const ITwinData = {};
  * @throws {RuntimeError} If all exports for the given iModel are Invalid
  * @throws {RuntimeError} If the iTwin API request is not successful
  */
-ITwinData.createTilesetFromIModelId = async function (options) {
-  let internalOptions = options;
-
-  if (typeof options === "string") {
-    // the old signature was (iModelId: string, options?: Cesium3DTileset.ConstructorOptions)
-    // grab the later arguments directly instead of in the params only for this special case
-    internalOptions = { iModelId: options, tilesetOptions: arguments[1] };
-    deprecationWarning(
-      "ITwinData.createTilesetFromIModelId",
-      "The arguments signature for ITwinData functions has changed in 1.132 in favor of a single options object. Please update your code. This fallback will be removed in 1.133",
-    );
-  }
-
-  const { iModelId, changesetId, tilesetOptions } = internalOptions;
+ITwinData.createTilesetFromIModelId = async function ({
+  iModelId,
+  changesetId,
+  tilesetOptions,
+}) {
   const { exports } = await ITwinPlatform.getExports(iModelId, changesetId);
 
   if (
@@ -100,6 +90,9 @@ ITwinData.createTilesetFromIModelId = async function (options) {
  * If the <code>type</code> or <code>rootDocument</code> are not provided this function
  * will first request the full metadata for the specified reality data to fill these values.
  *
+ * The <code>maximumScreenSpaceError</code> of the resulting tileset will default to 4,
+ * unless it is explicitly overridden with the given tileset options.
+ *
  * @experimental This feature is not final and is subject to change without Cesium's standard deprecation policy.
  *
  * @param {Object} options
@@ -107,30 +100,19 @@ ITwinData.createTilesetFromIModelId = async function (options) {
  * @param {string} options.realityDataId The id of the reality data to load
  * @param {ITwinPlatform.RealityDataType} [options.type] The type of this reality data
  * @param {string} [options.rootDocument] The path of the root document for this reality data
+ * @param {Cesium3DTileset.ConstructorOptions} [options.tilesetOptions] Object containing
+ * options to pass to the internally created {@link Cesium3DTileset}.
  * @returns {Promise<Cesium3DTileset>}
  *
  * @throws {RuntimeError} if the type of reality data is not supported by this function
  */
-ITwinData.createTilesetForRealityDataId = async function (options) {
-  let internalOptions = options;
-
-  if (typeof options === "string") {
-    // the old signature was (iTwinId: string, realityDataId: string, type: RealityDataType, rootDocument: string)
-    // grab the later arguments directly instead of in the params only for this special case
-    internalOptions = {
-      iTwinId: options,
-      realityDataId: arguments[1],
-      type: arguments[2],
-      rootDocument: arguments[3],
-    };
-    deprecationWarning(
-      "ITwinData.createTilesetFromIModelId",
-      "The arguments signature for ITwinData functions has changed in 1.132 in favor of a single options object. Please update your code. This fallback will be removed in 1.133",
-    );
-  }
-  const { iTwinId, realityDataId } = internalOptions;
-  let { type, rootDocument } = internalOptions;
-
+ITwinData.createTilesetForRealityDataId = async function ({
+  iTwinId,
+  realityDataId,
+  type,
+  rootDocument,
+  tilesetOptions,
+}) {
   //>>includeStart('debug', pragmas.debug);
   Check.typeOf.string("iTwinId", iTwinId);
   Check.typeOf.string("realityDataId", realityDataId);
@@ -168,9 +150,16 @@ ITwinData.createTilesetForRealityDataId = async function (options) {
     rootDocument,
   );
 
-  return Cesium3DTileset.fromUrl(tilesetAccessUrl, {
+  // The maximum screen space error was defined to default to 4 for
+  // reality data tilesets, because they did not show the expected
+  // amount of detail with the default value of 16. Values that are
+  // given in the tilesetOptions should still override that default.
+  const internalTilesetOptions = {
     maximumScreenSpaceError: 4,
-  });
+    ...tilesetOptions,
+  };
+
+  return Cesium3DTileset.fromUrl(tilesetAccessUrl, internalTilesetOptions);
 };
 
 /**
@@ -189,26 +178,12 @@ ITwinData.createTilesetForRealityDataId = async function (options) {
  *
  * @throws {RuntimeError} if the type of reality data is not supported by this function
  */
-ITwinData.createDataSourceForRealityDataId = async function (options) {
-  let internalOptions = options;
-
-  if (typeof options === "string") {
-    // the old signature was (iTwinId: string, realityDataId: string, type: RealityDataType, rootDocument: string)
-    // grab the later arguments directly instead of in the params only for this special case
-    internalOptions = {
-      iTwinId: options,
-      realityDataId: arguments[1],
-      type: arguments[2],
-      rootDocument: arguments[3],
-    };
-    deprecationWarning(
-      "ITwinData.createTilesetFromIModelId",
-      "The arguments signature for ITwinData functions has changed in 1.132 in favor of a single options object. Please update your code. This fallback will be removed in 1.133",
-    );
-  }
-  const { iTwinId, realityDataId } = internalOptions;
-  let { type, rootDocument } = internalOptions;
-
+ITwinData.createDataSourceForRealityDataId = async function ({
+  iTwinId,
+  realityDataId,
+  type,
+  rootDocument,
+}) {
   //>>includeStart('debug', pragmas.debug);
   Check.typeOf.string("iTwinId", iTwinId);
   Check.typeOf.string("realityDataId", realityDataId);
@@ -263,25 +238,11 @@ ITwinData.createDataSourceForRealityDataId = async function (options) {
  * @param {number} [options.limit=10000] number of items per page, must be between 1 and 10,000 inclusive
  * @returns {Promise<GeoJsonDataSource>}
  */
-ITwinData.loadGeospatialFeatures = async function (options) {
-  let internalOptions = options;
-
-  if (typeof options === "string") {
-    // the old signature was (iTwinId: string, collectionId: string, limit: number)
-    // grab the later arguments directly instead of in the params only for this special case
-    internalOptions = {
-      iTwinId: options,
-      collectionId: arguments[1],
-      limit: arguments[2],
-    };
-    deprecationWarning(
-      "ITwinData.createTilesetFromIModelId",
-      "The arguments signature for ITwinData functions has changed in 1.132 in favor of a single options object. Please update your code. This fallback will be removed in 1.133",
-    );
-  }
-
-  const { iTwinId, collectionId, limit } = internalOptions;
-
+ITwinData.loadGeospatialFeatures = async function ({
+  iTwinId,
+  collectionId,
+  limit,
+}) {
   //>>includeStart('debug', pragmas.debug);
   Check.typeOf.string("iTwinId", iTwinId);
   Check.typeOf.string("collectionId", collectionId);